fix: исправить все предупреждения clippy и unused_mut
- Исправлены предупреждения unused_mut в тестах (16 мест)
- Исправлены предупреждения assertions_on_constants (20 мест)
- Исправлены предупреждения identity_op (2 места)
- Исправлены предупреждения unnecessary_cast (1 место)
- Исправлены предупреждения double_comparisons (1 место)
- Исправлены предупреждения absurd_extreme_comparisons (1 место)

Все тесты проходят успешно, cargo clippy не выдаёт предупреждений

+++ src/testes/test_bag_system.rs
@@ -626,7 +626,7 @@ fn test_distribution_variance() {
     // Вычисляем дисперсию
     let variance: f32 = counts
         .iter()
-        .map(|&c| ((c as i32 - expected).pow(2)) as f32)
+        .map(|&c| ((c - expected).pow(2)) as f32)
         .sum::<f32>()
         / 7.0;
